refactor

+++ APIs/src/EpiServer.ContentGraph/Api/Filters/NumericFilterOperators.cs
@@ -9,37 +9,37 @@ namespace EPiServer.ContentGraph.Api.Filters
 
         public string Query => _query;
 
-        public NumericFilterOperators Boost(uint value)
+        public NumericFilterOperators Boost(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"boost: {value}" : $",boost: {value}";
             return this;
         }
-        public NumericFilterOperators Eq(long value)
+        public NumericFilterOperators Eq(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"eq: {value}" : $",eq: {value}";
             return this;
         }
-        public NumericFilterOperators NotEq(long value)
+        public NumericFilterOperators NotEq(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"notEq: {value}" : $",notEq: {value}";
             return this;
         }
-        public NumericFilterOperators Gt(long value)
+        public NumericFilterOperators Gt(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"gt: {value}" : $",gt: {value}";
             return this;
         }
-        public NumericFilterOperators Gte(long value)
+        public NumericFilterOperators Gte(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"gte: {value}" : $",gte: {value}";
             return this;
         }
-        public NumericFilterOperators Lt(long value)
+        public NumericFilterOperators Lt(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"lt: {value}" : $",lt: {value}";
             return this;
         }
-        public NumericFilterOperators Lte(long value)
+        public NumericFilterOperators Lte(int value)
         {
             _query += _query.IsNullOrEmpty() ? $"lte: {value}" : $",lte: {value}";
             return this;
@@ -49,12 +49,12 @@ namespace EPiServer.ContentGraph.Api.Filters
             _query += _query.IsNullOrEmpty() ? $"exist: {value.ToString().ToLower()}" : $",exist: {value.ToString().ToLower()}";
             return this;
         }
-        public NumericFilterOperators In(params long[] values)
+        public NumericFilterOperators In(params int[] values)
         {
             _query += _query.IsNullOrEmpty() ? $"in: [{string.Join(',',values)}]" : $",in: [{string.Join(',', values)}]";
             return this;
         }
-        public NumericFilterOperators NotIn(params long[] values)
+        public NumericFilterOperators NotIn(params int[] values)
         {
             _query += _query.IsNullOrEmpty() ? $"notIn: [{string.Join(',', values)}]" : $",notIn: [{string.Join(',', values)}]";
             return this;
